+++ tests/testthat/test_bit.R
@@ -15,11 +15,11 @@ test_that("checkBit", {
   expect_true(checkBit(xb, min.0 = 1))
   expect_true(checkBit(xb, min.1 = 1))
 
-  expect_string(checkBit(xb, len = 1), fixed = "length")
-  expect_string(checkBit(xb, min.len = 3), fixed = ">=")
-  expect_string(checkBit(xb, max.len = 1), fixed = "<=")
-  expect_string(checkBit(xb, min.0 = 2), fixed = "'0'")
-  expect_string(checkBit(xb, min.1 = 2), fixed = "'1'")
+  expect_error(assertBit(xb, len = 1), regexp = "length")
+  expect_error(assertBit(xb, min.len = 3), regexp = ">=")
+  expect_error(assertBit(xb, max.len = 1), regexp = "<=")
+  expect_error(assertBit(xb, min.0 = 2), regexp = "'0'")
+  expect_error(assertBit(xb, min.1 = 2), regexp = "'1'")
 
   expect_error(checkBit(xb, len = NA), "missing")
   expect_error(checkBit(xb, min.len = NA), "missing")