+++ tests/testthat/test_checkTibble.R
@@ -2,9 +2,10 @@ context("checkTibble")
 
 test_that("checkTibble", {
   skip_if_not_installed("tibble")
-  library(tibble)
+  requireNamespace("tibble")
 
-  x = as_tibble(iris)
+  x = tibble::as_tibble(iris)
+  expect_succ_all("DataFrame", x)
   expect_succ_all("Tibble", x)
   expect_fail_all("Tibble", iris)