+++ R/checkDataTable.R
@@ -26,9 +26,9 @@ checkDataTable = function(x, key = NULL, index = NULL, types = character(0L), an
 }
 
 checkDataTableProps = function(x, key = NULL, index = NULL) {
+  if (!requireNamespace("data.table", quietly = TRUE))
+    stop("Install package 'data.table' to perform checks of data tables")
   if (!is.null(x)) {
-    if (!requireNamespace("data.table", quietly = TRUE))
-      stop("Install 'data.table' to perform checks of data tables")
     if (!data.table::is.data.table(x))
       return("Must be a data.table")
     if (!is.null(key)) {