+++ R/checkTibble.R
@@ -15,13 +15,15 @@
 #' testTibble(x)
 #' testTibble(x, nrow = 150, any.missing = FALSE)
 checkTibble = function(x, types = character(0L), any.missing = TRUE, all.missing = TRUE, min.rows = NULL, min.cols = NULL, nrows = NULL, ncols = NULL, row.names = NULL, col.names = NULL, null.ok = FALSE) {
+  if (!requireNamespace("tibble", quietly = TRUE))
+    stop("Install package 'tibble' to perform checks of tibbles")
   qassert(null.ok, "B1")
   if (is.null(x)) {
     if (null.ok)
       return(TRUE)
     return("Must be a tibble, not 'NULL'")
   }
-  if (!inherits(x, "tbl_df"))
+  if (!tibble::is_tibble(x))
     return(paste0("Must be a tibble", if (null.ok) " (or 'NULL')" else "", sprintf(", not %s", guessType(x))))
   checkDataFrame(x, types, any.missing, all.missing, min.rows, min.cols, nrows, ncols, row.names, col.names, null.ok)
 }