+++ NEWS.md
@@ -1,4 +1,5 @@
 # Version 1.8.4
+* `*Names` with `type="unnamed"` now works with `NULL`.
 * Fixed possible protection stack imbalance as reported by `rchk`.
 
 # Version 1.8.3