+++ NEWS.md
@@ -2,6 +2,7 @@
 * `*Names` with `type="unnamed"` now works with `NULL`.
 * Fixed possible protection stack imbalance as reported by `rchk`.
 * New argument `must.include` for `*Names`.
+* New functions to test bit vectors implemented in package `bit`
 
 # Version 1.8.3
 * New argument `sorted` (defaults to `FALSE`) for `*Integer`, `*Integerish` and